[chore] Move capabilities consumers in a separate package
This is to prepare to move "pipelines" as public, see https://github.com/open-telemetry/opentelemetry-collector/issues/5564
